Guia 13

+++ Backend 1 - Intro a Java/Guia 13 - Manejo de Excepciones/Ejercicio Integrador/G13_EjIntegrador_IronMan/src/entities/armadura/piezas/Bota.java
@@ -14,7 +14,7 @@ public class Bota {
         propulsor = p;
     }
 
-    public float usar(int intensidad, int tiempo){
+    public float usar(int intensidad, float tiempo){
         return propulsor.usar(intensidad,tiempo);
     }
