+++ Backend 1 - Intro a Java/Guia 13 - Manejo de Excepciones/Ejercicio Integrador/G13_EjIntegrador_IronMan/src/entities/armadura/piezas/dispositivos/Propulsor.java
@@ -15,7 +15,7 @@ public class Propulsor extends Dispositivo {
         this.consumo = consumo;
     }
 
-    public float usar(int intensidad, int tiempo){
+    public float usar(int intensidad, float tiempo){
         return (consumo * tiempo) * intensidad;
     }
